safeguards refactoring & reset safeguard api & typos and clear variable naming
